moved layer util functions to esriLayerUtils
removed esriMapUtils.addLayerToMap()
other map util functions remain in esriMapUtils
